Revert "- Remove 'struct vmmeter' from 'struct pcpu', leaving only global vmmeter"
This reverts commit 21ead51d79e6c9c2eb017eff84a3c8ae660a79b5.

+++ usr.bin/top/machine.c
@@ -516,7 +516,7 @@ get_system_info(struct system_info *si)
 		static int swapavail = 0;
 		static int swapfree = 0;
 		static long bufspace = 0;
-		static uint64_t nspgsin, nspgsout;
+		static int nspgsin, nspgsout;
 
 		GETSYSCTL("vfs.bufspace", bufspace);
 		GETSYSCTL("vm.stats.vm.v_active_count", memory_stats[0]);
